add double tap gesture to play pause the video
