+++ feature/player/src/main/java/dev/anilbeesetti/nextplayer/feature/player/buttons/PlayerButton.kt
@@ -1,0 +1,49 @@
+package dev.anilbeesetti.nextplayer.feature.player.buttons
+
+import androidx.compose.foundation.combinedClickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PlayerButton(
+    modifier: Modifier = Modifier,
+    shape: Shape = CircleShape,
+    contentPadding: PaddingValues = PaddingValues(8.dp),
+    containerColor: Color = MaterialTheme.colorScheme.secondaryContainer,
+    contentColor: Color = MaterialTheme.colorScheme.onSecondaryContainer,
+    onClick: () -> Unit,
+    onLongClick: (() -> Unit)? = null,
+    content: @Composable () -> Unit,
+) {
+    Surface(
+        modifier = modifier
+            .clip(shape)
+            .combinedClickable(
+                role = Role.Button,
+                onClick = onClick,
+                onLongClick = onLongClick,
+            ),
+        shape = shape,
+        color = containerColor,
+        contentColor = contentColor,
+    ) {
+        Box(
+            modifier = modifier.padding(contentPadding),
+            contentAlignment = Alignment.Center,
+        ) {
+            content()
+        }
+    }
+}
