+++ feature/player/src/main/java/dev/anilbeesetti/nextplayer/feature/player/buttons/RepeatButton.kt
@@ -1,0 +1,41 @@
+package dev.anilbeesetti.nextplayer.feature.player.buttons
+
+import androidx.compose.material3.Icon
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.media3.common.Player
+import androidx.media3.ui.compose.state.rememberRepeatButtonState
+import dev.anilbeesetti.nextplayer.core.ui.R as coreUiR
+
+@Composable
+fun LoopButton(player: Player, modifier: Modifier = Modifier) {
+    val state = rememberRepeatButtonState(player)
+
+    PlayerButton(modifier = modifier, onClick = state::onClick) {
+        Icon(
+            painter = repeatModeIconPainter(state.repeatModeState),
+            contentDescription = repeatModeContentDescription(state.repeatModeState),
+        )
+    }
+}
+
+@Composable
+private fun repeatModeIconPainter(repeatMode: @Player.RepeatMode Int): Painter {
+    return when (repeatMode) {
+        Player.REPEAT_MODE_OFF -> painterResource(coreUiR.drawable.ic_loop_off)
+        Player.REPEAT_MODE_ONE -> painterResource(coreUiR.drawable.ic_loop_one)
+        else -> painterResource(coreUiR.drawable.ic_loop_all)
+    }
+}
+
+@Composable
+private fun repeatModeContentDescription(repeatMode: @Player.RepeatMode Int): String {
+    return when (repeatMode) {
+        Player.REPEAT_MODE_OFF -> stringResource(coreUiR.string.loop_mode_off)
+        Player.REPEAT_MODE_ONE -> stringResource(coreUiR.string.loop_mode_one)
+        else -> stringResource(coreUiR.string.loop_mode_all)
+    }
+}
